Change from unique constraint to primary key constraint for edge tables because of the following:
"Adding a primary key will automatically create a unique B-tree index on the column or group of columns listed in the primary key, and will force the column(s) to be marked NOT NULL.

A table can have at most one primary key. (There can be any number of unique and not-null constraints, which are functionally almost the same thing, but only one can be identified as the primary key.) Relational database theory dictates that every table must have a primary key. This rule is not enforced by PostgreSQL, but it is usually best to follow it."

from https://www.postgresql.org/docs/9.4/ddl-constraints.html

so even though postgres doesn't enforce it, may as well do the right thing now because we may run into issues with other db dialects in the future.