Add a simple v0.01 cache
Because of the chatty nature of things, it involves lots of back and forth to the db

Right now, we're caching loading of individual "nodes": edge_data and nodes like events, event group etc

this is a short term thing to test ent-rsvp

Long term, we need to rewrite this entire layer and break it down into 2 different caches:
1/ edge data which is always cached in memory since it doesn't change often.
2/ node data which needs to be cached per request in case we run into it again because of the cyclical nature of data and then selectively cached in memory/redis/etc as needed

had to change the way we load data for each node to use a MapScan() to extract the data from sqlx.Rows and then store that Map and then fill the data ent from the map
Doing this because I think it's the best long term way of getting data once we start getting data from different caches. If that doesn't work, we'll need a Clone() method on each ent so that what's stored in the cache after StructScan isn't what's returned to each subsequent call since we're caching pointers/sending pointers down and want to make sure that we send a clean copy

I think this is probably better because it removes the need for relect in this path. Slowly but surely we'll remove the lazy parts that had reflect to start