change dataEntity interface to add a DBFields() method instead of FillFromMap().
Instead of using reflection on each item to get the columns to fetch, we're going to use the keys returned from this method to indicate the columns from the database that are fetchable.

This diff doesn't change the columns yet. It uses DBFields() and the value of each key (the function returned) to go through and set the value of each field similarly to what fillFromMap() was doing.

A subsequent change will change the columns code. And a further change will change the generated code to have the right code