actions with privacy
this starts the process of confirming that the privacy framework can work with other things.

Doesn't work well with actual actions since I have to go back and fix how things are currently structured for real live actions to fit this but that'll happen in the future soon.

The one downside now is that the ent can be nil in the case of a create action and now all rules have to accoutn for that? that doesn't sound like fun...