make dataOperation in ent public + make changeset API simpler
A changeset returns an executor (need a better name) which is responsible for handling order operations and dependencies which are needed to perform a mutation.
Right now, the operations we support are nodes and edges as part of a single change to one single node and we can just use a ordered list to handle that.

However, once we add dependencies across nodes e.g. create a user and contacts at the same time and need the user id created to be stored in each contact object, things get more complicated quickly.

Having the public API be simple and let each case handle the complex things makes things a lot easier.

And once we add critical observers, we can chain all of these correctly and make sure they all happen within the same transaction and the public API is none the wiser as to what complicated things are happening in there.

Instead of using the AugmentWithPlaceHolder method API, we can just create dependencies within the changeset e.g. edge operations depend on the create operation
or changesets that depend on others e.g. creating a user and contacts with edges. initially, the only dependencies will be the containing mutation but eventually it could be even more complicated

step 1: we create the user
step 2: we create the contact and get the user id and store in the contact object because it depends on the user id created in step 1
step 2b: we create add an edge from user to something else. this can be run at the same time/within the same pass as step 2

and go on from there and see where the music takes us.

ExistingEnt() and EntConfig() I think are no longer needed in a changeset but keeping that anyways.