have Executor interface handle the resolving of temporary placeholder ids
now we can have different executors depending on how complicated things are

Because we're depending on an interface, we can change complexity/implementation over time and have it be constrained to specific scenarios.

When we have multiple changesets involved at the same time, we can use a more complicated Executor which has a mapping of placeholder value to id and figures out dependencies between changesets or we can rewrite different things for tests etc

To make it easier for EdgeOperation to handle when there's placeholders or not (since it's no longer a constant), made all the fields private and made it easier for the edge to track what's going on. Completely changed the API so EdgeOperation has completely control over the values (package private)