change action and builder generation so tha an action always has one builder
with the refactor earlier to add MutationBuilder as part of changesets, couldn't have a changeset with multiple builders so needed to collapse EdgeGroupMutationBuilder into MutationBuilders that have status groups.

And this means that if we have triggers associated with these actions, they can get the right builder and make whatever other ancillary changes they want to do which was the original reason these actions had multiple builders.
On the flip side other actions which don't use these fields directly can now set these fields in their triggers which is good.