add action.IsRequiredField
add test for it. we don't have any other tests for internal/action so this is first one.
Refactor some of the logic surrounding parse_helper and tests.
add caching by adding testingutils.sync.RunOnce so that tests can run slightly faster and not have to re-parse a lot of the files repeatedly.