add ent.Edge interface (#14)
add ent.Edge interface

makes it cleaner and types ent.EdgeMap "correctly". makes it clearer what the API for GetEdges() is. The fact that we load the packages when we parse the AST should enforce the right APIs.
 
changes ent.AssociationEdgeGroup's EdgeGroups field to not be ent.EdgeMap but ent.AssocEdgeMap