simplify ViewerContext interface
Long meant to do this.

Makes it easier for new things to start with the framework (will eventually have an init command that creates the default ViewerContext for a project) and removes concepts that don't need to be included with every instance of ViewerContext e.g. omniscient.

Removes HasIdentity method from ViewerContext and just provides function in viewer package that does the same job

Also changes all instances of "viewer viewer.ViewerContext" to "v viewer.ViewerContext" so that viewer package isn't hidden in methods since we may end up with more package-level functions later