add enttype.FieldType as the first pass of "overhauling" what we do for types
uses types.TypeString() to get the underlying string definition of the type and uses that to map to strings, ints, floats, times etc with top level support for the different types and getting the graphql definition as needed

This changes only the custom rendering of params and return types to use this for now. Will change other places most notably fields to use new API soon/next.

Most of the code in enttype/field_types.go is copied from field/field_types.go with the expectation that this will get combined shortly