working version of input types for mutations
plus a flag with the option to disable the input type for whatever reason.

TBD: if we are disabling input type, should we also provide option to disable Response type? seems like no because graphql doesn't have void type and it's easier to be consistent otherwise disabling it when there's no return is an error and it's yet another edge case to consider