add tests for types
this also fixes a bunch of random things that were broken because of lack of test coverage.

This should also hopefully make a bunch of things clearer.

PointerType, NamedType and fieldWithActualType can still be made clearer and cleaned up but for now this works and we have tests to confirm expected behavior