add new request and auth packages
request package provides a Register function that can be used to Register http handlers that get "integrated" with whatever router is being used (defaults to mux for now)
this makes it so that other handlers can be added to the framework e.g. cors handling, memory, logging etc and not have to interact with the codegen process

auth package also provides a Register function that can be used to Register auth handlers which take the response/request and then returns a viewer

auth package registers 1 handler with request package and then goes through the list of its auth handlers to eventually return a viewer to be used with the system.

If no viewer is returned by any of the handlers (or if no handlers), we default to the logged out viewer context