Add default implementation of phone number/pin auth
This isn't perfect by any means but gets simple auth that plays well with the framework and provides a starting point to test thigns.
Right now, it validates any 6 digit PIN as valid. Need to hook redis and/or memory validation in there

Adds PhonePinAuth struct which implements auth.Auth which provides 2 main APIs:
* Authenticate() which takes a context and a phone number/pin combo and returns an authed identity with a viewercontext and a token
This should be called from the GraphQL "auth" Mutation or REST API handler as a response to a user action
* AuthViewer which takes http.ResponseWriter and http.Request and returns a ViewerContext.
This implements the auth.Auth interface.

Client needs to call `auth.Register("phone_pin_auth", phonenumber.NewPhonePinAuth(...))` to register this auth handler to deal with requests as needed

Adds a nullable unique phone number to test_schema User model and adds an accessor to get IDs of unique/indexed fields. This adds LoadUserIDFromEmailAddress and LoadUserIDFromPhoneNumber to models.User