Add default implementation of email/password auth
Add password to user. Change FieldInfo to know the path and generate ValidateEmailPassword function which can then be passed to EmailPasswordAuth to validate email/password combo

Optimized for now exposing the password out of the package (it's currently exposed but will hopefully soon be a private field)

Created SharedJwtAuth to share logic between EmailPasswordAuth and PhonePinAuth. Put it in an internal package to not have AuthFromID be a public API. we'll see how annoying that ends up being

This makes EmailPasswordAuth and PhonePinAuth slightly more annoying to implement but think it's worth it for now...

and other changes