adds default auth0 integration
requires issue and audience, verifies they're valid based on token passed along from auth0

went through the tutorial at https://auth0.com/docs/quickstart/backend/golang/01-authorization and adapted it to be as generic as possible and hopefully giving clients as little work as possible

right now we cache the certificate "forever" (as long as the go server is up) since it doesn't seem like it changes. In an ideal world, this is smarter and caches for a more sensible period. Cache theoretically shouldn't grow crazily since in most cases we should only be doing this for one issue as opposed to a lot

This should ideally eventually be its own github package but putting everything "core" in one big package for now until it's really good and can be improved as needed

Also fix a bug with private fields and trying to struct scan from them