Add initPhoneAuth command to gent
Provides a way to get quick or sensible defaults for phone number/pin account create and/or registration.

It doesn't handle the account creation step (yet) because of what fields may be needed via actions.

It still depends on the user creating a manual function to create the user (and validate the pin) or adding a mutation only field to the action and checking it there. Will eventually have better
support here.

Run as follows: `gent initPhoneAuth -p models/configs -n User -f PhoneNumber -v testschemaviewer.NewViewerContext --force`
If -v isn't provided, it creates a local viewer context to be used.

Because we can't call functions in global scope (or whatever the language go uses here), developer has to copy the one-liner into an init() function of some sort and call register.
Good thing about everything going into this init() function is that it's clear what magic is going on there

Code doesn't have the "Code generated by ... header" so developer can change as needed: adding and/or removing functionality