add way to extend auth token.
this is added to both PhonePinAuth and EmailPasswordAuth via SharedJwtAuth

It doesn't extend tokens by "default"

As of right now, the auth APIs added by `gent init...` add a 10 minute window before the token is expired as the default. That's subject to change.
It's also possible to not do that and send a new token back with each request but that has to be done manually by adding a new middleware at the end. I don't believe we have a way to guarantee that right now actually...

adds some tests and calls `gent codegen -p models/configs` to generate the graphql code needed from running this