convert ent.GenLoadEdgesByType to new API
We return a channel for concurrent APIs instead of what we previously had with a WaitGroup and a channel.
code is a lot cleaner and easier to use

Only issue I can see is that we don't need go func() in these return types here and that may be confusing?