convert ent.GenLoadUniqueNodeByType to new API format which returns a channel
We return a channel for concurrent APIs instead of what we previously had with a WaitGroup and a channel.

The sole callsite: Event.GenCreator() is still complicated however clients are a lot simpler. see models_test.go

In an ideal world combo of GenCreator|GenLoadUniqueNodeByType will be simpler but it's not immediately obvious how so keeping this as is.
If there was a SingleEntLoader interface like
```golang
type SingleEntLoader interface {
	GetEnt() dataEntity
	GetError() error
}```
and GenLoadUniqueNodeByType returned a channel to that, that could work but would need to pass it in so hmmmm