make Loader used by all public APIs which load ents
Also changes return values when not successful from zero value of object to nil. It's cleaner and less prone to subtle errors down the stack. Now, it clearly blows up if someone is doing the wrong thing which is the right thing in the long run

Removes reflection in the loading process. Each loader responsible for setting the viewer as needed. And no need for SetZeroVal either since we just set the return value to nil

Plus a bunch of other cleanup with renaming and other things.