have LoadNodeFromEmail methods go through loadNodeLoader instead of loadNodesLoader
it makes the result consistent with what we get from using QueryRow vs rows.Next()/keep scanning

when there's one result we're looking for, we use loadNodeLoader and can more easily test for that.

When we're fetching potenitally multiple resutls, we go through loadNodesLoader and no result isn't an error and is consistent with dbsql/sqlx APIs

This also simplifes loadNodesLoader since we don't have to check for this lack of error and return an error. Also simplifies edge_data.go