add base field classes needed for API.
API for clients a lot easier to use and we get strongly typed options

Still need to spend time polishing the API but I think the template we have here works and will work for both custom and non-custom types.

I think the corresponding factory method with each type is the easier approach for the calling clients but theoretically subject to change

Need to look at what API is more common in typescript for options. String (the only one filled) here has a mix of options and chainable methods.

Need to also clean up some of the code duplication in each of the factory methods.