typescript integration beginning (#27)
Typescript schema. With Fields!

Given a typescript schema and a command to run tsent codegen, ingest the schema and generate the database similar to what is done for gent codegen.

Unlike go, I've learned my lesson and not trying to do an AST here. I'm generating a new file using read_schema.tmpl and getting a JSON representation of the schema that the user provided and using that as input to generate the NodeMap|NodeData etc.

Will still eventually need AST when I get to custom objects and the like but not there yet.

Also change the golang AST code to go from AST/package -> field inputs -> Schema

This will slowly transition things over as new functionality is added for typescript and eventually the only logic in AST land will be going from AST parsing to inputs.

Chose to build the CLI in golang since already using that. Therefore using ts-node to run the generated typescript code.

In addition to Fields, schema also supports a concept called a pattern. Allows reusing functionality amongst different schema. First (and probably most common) pattern is the Node schema which adds id, created_at, updated_at fields to each node. Added a BaseEntSchema class to make it easier to use. Also have a Timestamps schema that Node uses.