add Edges to typescript schema
This adds AssocEdges and AssocEdgeGroups as configurable edges to the typescript schema.

Doesn't add foreign key edge (similar to go) since we can infer it from the foreignkey field property (and generate the foreign key and field edge from it)

Doesn't add field edge (will change go to do this) because it should be on field for a better API (similarly to foreign key) and then we can generate the field edge from it

There's a potential third case for field edge where there's a field that's a not a foreign key or an edge for either legacy or performance reasons (or something non-obvious) at the moment and we want to indicate what the object is. Hopefully, the API can just change to ["User", ""] and it should just work when we get there or ["User"] in typescript since it's more flexible.