don't hardcode database connection
1/ takes it from connection string if it exists
2/ takes it from config/database.yml next

right now this is done implicitly when DB.getInstance() is called down the line.
In an ideal world, we probably init this once but we can deal with this later.

this is dependent on seemingly having a package.json in the root of the project and reading relative from that. not 100% sure this is accurate but that seems to be what works. will need to test this in multiple environments to be sure of this.