add viewer and privacy checks to ents (#30)
This adds a Viewer interface to ent and passes it to all APIs

Adds Privacy Policy, Privacy Policy rules and everything needed to apply privacy to an ent

Each ent defaults to not being visible to anyone.

Split up ent into 2 classes e.g. UserBase and User. UserBase is never going to be edited by the user while User is a one time generation that can then be continually edited by the user. Initial thing expected to be edited is the privacy policy for the ent. custom ent methods and other user logic will eventually be added here.

Uses generics. Typescript is nice :)