change structure of generated code to make it easier to use
Core issues we care about:
* We want to have different files here so as to not have generated code mix with user|non-generated code.
* We should be able to regenerate the generated code file as many times as possible without affecting/mixing with user-code
* We should be able to add new functionality to the schema and have the existing user-code class keep working.
* developer should be able to extend the base class to add new functionality. easiest/most-obvious is override privacy policy for the ent. we want the privacy policy that the developer provides/edits to keep working

In go, we're doing this by breaking up the class into multiple files and that works well since classes can be broken up into multiple files and even private functionality works across classes.
Typescript doesn't support breaking up a class into multiple files so not an option.

Typescript's traits are weird and don't really work for what we want because support for static members not great so far.

So 2 classes: base and derived seems like the best approach (what we currently have).

Have to modify what we currently have to support all the cases above.

For example, if we add an indexed field or a unique email address and a
`UserBase.loadFromEmailAddress(viewer: Viewer, emailAddress: string): Promise<UserBase | null>`

method is generated, nothing from User should need to change and we should be able to call it like `const user = await User.loadfromEmailAddress(new LoggedoutViewer(), "test@email.com");` and that should work, return an instance of `User` not 'UserBase` and new functionality added in the `User` class comes with it.

spent a bunch of time looking into the issue and this seems like a known typescript isssue. https://github.com/microsoft/TypeScript/issues/5863

Ended up with an approach that works but isn't perfect. Most annoying are
1/ the base classes will need to be aware of the generated ones but typescript has no import cycle issues so it's "fine".
2/ the base class is not abstract and could theoretically be instantiated on their own. Will have to be enforced through culture, convention, commit hooks or other means