resolve placeholders in operations so that the right ids can be written in dependent operations (#46)
this practically allows us to write edges in the builder at creation e.g.

``` ts
// dany and sam exist above...
  let action = CreateUserAction.create(loggedOutViewer, {
    firstName: "Jon",
    lastName: "Snow",
    emailAddress: randomEmail(),
  });
  let t = new Date();
  t.setTime(t.getTime() + 86400);
  action.builder.addFriend(dany).addFriendID(sam.id, {
    time: t,
  });
  const jon = await action.saveX();
```

allows us to create a user, add 2 friends and change the timestamp of the 2nd friendship to a custom time all in the same transaction.

The logic for adding friends waits for the user to be created and then gets the newly created id from that and uses that when writing the edge.

this addFriend logic should eventually be in a trigger (coming soon) but for now this will be done manually. Also, the good thing is any developer *should* always be able to cobble one off like this if for some reason they don't want to use triggers although the full action flow with triggers/observers etc always preferred.