recognize id fields  (#47)
* change type of ID types from string to ID.

this changes it in generated ent, builder, action

also includes a hack in builder/action that checks for "ID" specifically and makes sure it's imported. Will be done generically eventually when we have other custom types to include here.

* generate code to add inverse edge to idfield if an inverse edge exists when the field is written

this also fixes a bug (also in golang!) where the wrong idtype wasn't being written for the inbound edge.

we were writing the idType for the source e.g event  instead of the destination e.g. user for the following (current event.ts schema in simple):

```ts
export default class Event extends BaseEntSchema implements Schema {
  fields: Field[] = [
    // TODO this should be an id type...
    // we should warn when we see an "ID/id/Id" field as non-id type and ask if they wanna change it
    StringType({
      name: "creatorID",
      fieldEdge: ["User", "createdEvents"],
      storageKey: "user_id",
    }),
  ];
 }

```

This also fixes another bug in golang that I found where we were trying to validate a builder and casting to string

We also need to change the type in builder and action to add Builder<{{Node}}> but will do that when we support triggers (coming soon)

* fix comment