kill required fields in generated builders
don't need them since we require the input at once so we can depend on typescript's strong typing to handle it.

if we don't have typescript or strong typing of the sort, we'd need to change this.

should probably also change golang to use something like this instead of what we currently have with the setters.
the problem is that go is not as flexible with field types the way typescript is with its interfaces