make it so that we can add edges to edges that are also being created in changesets
e.g. a trigger that adds a contact and adds an edge to that contact

have to change how executor works to not only understand dependencies but to run operations that run rows first
(will add full suite of tests next)

this changes the approach from what we were doing in golang to inline all the logic within the Executor and have the outer executor be responsible for reordering the operators and resolving the value

instead of building depgraph or something like it, quick google search found https://github.com/datavis-tech/graph-data-structure which seems to do what I want so using that.