typescript: add triggers to actions (#50)
add triggers to actions

this makes it possible to batch side effects to each action that can be run in the same transaction as the action.

these triggers don't actually perform any writes in them. they use the `builder` to build up modifications to the existing changeset or return additional changesets that should also be run as part of the mutation/action.

At some point in the future when we're super fancy, we can detect writes happening in triggers and respond to that in some way. For now, nothing's happening here and it's not as performant as it could be.

It's still TBD on exactly when triggers should run. In go, they happen before validation. here, some triggers run before validation (if they don't return a promise), and some run parallel to validation. This inconsistency is not going to work long term and probably need to change and figure out best place to do so. This is where feedback and getting real world usecases again will be helpful.

For example, every time an account is created, marking the user as "UNVERIFIED"

```ts
  // assuming an accountStatus field on the user 
  triggers = [
    {
      changeset: (builder: UserBuilder): void => {
        builder.updateInput({
          accountStatus: "UNVERIFIED",
        });
      },
    },
```

or more complicated, to also create a contact for the user when the user's account is created assuming a contact management app where the user always has contacts including themselves as one:

```ts
export default class CreateUserAction extends CreateUserActionBase {
  triggers = [
    {
      changeset: (builder: UserBuilder): void => {
        builder.updateInput({
          accountStatus: "UNVERIFIED",
          // not needed because we have serverDefault but can also set it here.
          emailVerified: false,
        });
      },
    },
    {
      // also create a contact for self when creating user
      changeset: (builder: UserBuilder): Promise<Changeset<Contact>> => {
        let input = builder.getInput();
        let action = CreateContactAction.create(this.builder.viewer, {
          firstName: input.firstName!,
          lastName: input.lastName!,
          emailAddress: input.emailAddress!,
          userID: builder,
        });

        builder.addSelfContact(action.builder);
        return action.changeset();
      },
    },
  ];
}
```

or to auto-join a user to default channels in a workspace ala slack:

``` ts 
action.triggers = [
    {
      changeset: async (
        builder: SimpleBuilder<Group>,
      ): Promise<Changeset<any>[]> => {
        let [userInfo, autoJoinChannels, invitee] = await Promise.all([
          fetchUserName(),
          getAutoJoinChannels(),
          getInvitee(builder.viewer),
        ]);
        userAction = new UserAction(
          builder.viewer,
          new Map([
            ["FirstName", userInfo.firstName],
            ["LastName", userInfo.lastName],
            ["EmailAddress", userInfo.emailAddress],
          ]),
          WriteOperation.Insert,
        );
        for (let channel of autoJoinChannels) {
          builder.orchestrator.addOutboundEdge(
            channel.id,
            "channelMember",
            "Channel",
          );
        }
      return userAction.changeset();
  },
}];
```