TypeScript decorators to support custom fields (#65)
This is the beginning of the high level work to support custom queries and fields in ents exposed in `GraphQL`.

with the approach taken in https://github.com/lolopinto/ent/pull/63, we need a way to integrate custom fields into the generated graphql schema. 

The approach I'm taking here is to use decorators which are "ignored" at run time and only evaluated during the codegen process to gather information about any custom properties or methods to be integrated into the graphql generation process.

The plan is to then take the information received and then generate graphql accessors/properties for these methods. Future issues I see here is that there doesn't seem to be an obvious way to have access to the imported files/paths etc so if there's external dependencies, we'd need to make sure we import them correctly. Haven't solved that problem yet and may need to end up with complicated ways to solve that e.g. using the TypeScript compiler.

Will deal with that at some point.

Next step is to take the output of this and do the work to integrate with #63