GraphQL integration for TypeScript (#63)
history/context is based on research from https://github.com/lolopinto/ent/issues/59

This is using `graphql-js` directly which is the reference implementation for graphql that a lot of the JavaScript/TypeScript implementations use.

Since code is generated here, seems to make sense to go directly to the source.

Will probably end up with some small wrapper around graphql-js to make it easier to use but for now it's fine and we're going straight to the source.

This provides the reference implementations based on the ent schema, generates root queries for each object, generates a mutation for each action, generates a query and mutation root type and for now manually just puts it all together in `src/graphql/index.ts`. It also generates a `schema.gql` file based on the schema and uses `printSchema` of the `graphql-js` library to do it as opposed to manually generating the schema file as we do for go.

For mutations, it generates an input type for each mutation based on the [`relay`](https://relay.dev/) way of doing things. 

Will eventually have options for some of these things but for now it's not as flexible