GraphQL tests (#64)
this writes tests for the generated graphql schema to make sure it works

has tests for the following:

- user type
- event type 
- contact type
- user mutations (create, edit, delete) and checks for permissions

this makes sure the generated code does what it's supposed to do 

this also includes a generalizable graphql testing framework that helps to build queries and makes it easier to use graphql. e.g. in `user_type.test.ts`. It'll take parts of a query such as 

```    ["id", user.id],
    ["firstName", user.firstName],
    ["lastName", user.lastName],
    ["emailAddress", user.emailAddress],
    ["accountStatus", user.accountStatus],
    ["selfContact.id", selfContact.id],
    ["selfContact.firstName", selfContact.firstName],
    ["selfContact.lastName", selfContact.lastName],
    ["selfContact.emailAddress", selfContact.emailAddress],
    ["selfContact.user.id", user.id],
    ["friends[0].id", user3.id],
    ["friends[0].firstName", user3.firstName],
    ["friends[0].lastName", user3.lastName],
    ["friends[1].id", user2.id],
    ["friends[1].firstName", user2.firstName],
    ["friends[1].lastName", user2.lastName],
```

and generates a query of the form `user(id...) {id, firstName, lastName, emailAddress, accountStatus,selfContact { id, firstName, lastName, emailAddress, user {id}}, friends {id,firstName, lastName}}}` (not tested in this case)

this makes it easier to write graphql tests against the schema, ensure that the tests we're writing are similar to what the client does when it hits graphql and can see if there's any issues anywhere along the pipe even in serialization.