some small fixes to phone number type.
found out that the default validation is not as strict as expected

e.g. "1" is an invalid US number but "4152" is not

Looking at the code, it says to call isPossible() by default and to not call isValid() because that's a whole mess

so, changing the behavior so the default behavior calls isPossible() (and doesn't allow obviously invalid numbers) while providing an option to disable that. Also provides an option to call isValid() and for custom validation