TS:  Testing improvements (#137)
this makes it easier to test querying/writing to db in the core framework

ideally, we'd use sqlite as an in-memory db here but we'd still need to write SQL to generate the table schema and all the code currently used to generate schema is in python (auto_schema) and didn't seem worth it to call into a different language for this or to re-invent that wheel.

Instead, I chose to parse the sql and do in-memory management of a schema-less database just to confirm that basic things are working.

Will be especially useful when it's EntQuery time since we'll need to confirm that slicing, ordering etc works and easiest to just load a lot of data in memory and apply the sql on it as needed.

This changes everything currently using QueryRecorder to use the new format and no more jest mocking of certain functions to fake the data. Now, data is just written and retrieved like normal.

Only 2 lines necessary to mock are:

jest.mock("pg");
QueryRecorder.mockPool(Pool);
and normal INSERT, UPDATE, SELECT, DELETE statements are used.

Current limitations:

only supports prepared statements for now. doesn't work for inline values yet
only supports postgres prepared statements e.g. '$1', and not mysql '?'
doesn't support all select options e.g. no GROUP BY, WITH, DISTINCT, HAVING etc as those are not currently used in ent. Can be added as needed later.
Uses node-sql-parser to parse the SQL and is subject to the current limitations of that library e.g. taozhi8833998/node-sql-parser#470 and ON CONFLICT DO UPDATE not being parsed at the moment.

Also moves the query predicates from query to clause internally to prepare for EntQuery.