Generate EntQuery in generated ent code (#153)
This builds on #142.

It generates an entquery for each assoc edge referenced in the schema.

It generates 2 classes for each query. A base class that will be autogenerated each time. Like actions and ents, this is owned by the ent layer and can keep expanding. And then a subclass which will be user code. The subclass doesn't currently do much but will soon support per edge privacy and that's where the privacyPolicy will go.

per edge privacy allows us to control who can see each edge in the graph. for example, you can imagine only the user being able to see who they have blocked but everyone can see who the user's friends are or who they follow. etc

another benefit of the base class being separate is that as more edges get added to other schemas, we can keep modifying those to add accessors to those to have chained queries working