publish to github package repo  (#176)
so that we can just grant credentials in github and not also need a docker account

this follows the steps at https://docs.github.com/en/free-pro-team@latest/packages/managing-container-images-with-github-container-registry/pushing-and-pulling-docker-images#authenticating-to-github-container-registry

every new version is currently manual. in the future, we could use [github actions](https://docs.github.com/en/free-pro-team@latest/actions/guides/publishing-docker-images#publishing-images-to-github-packages) to auto-publish but not yet.