Custom entquery edge (#213)
This allows custom data in edges that are reflected throughout the system.

We generate a CustomEdge class for each AssocEdge that the developer can add any code to.
Developer can then also add the graphql decorators to that edge so that the custom fields are reflected in the generated graphql similar to what we for custom methods in ent or top level custom queries and mutations.

An example is dietary restrictions on an rsvp to an event. Since this is "edge data" e.g. data associated with the connection btw an event and the user who's rsvping, it makes the most sense to store this on the edge e.g. either directly in the data field or in a new object with the id stored in the data field

And then the developer, can then fetch that data and pass it all the way down through the system so that it shows up in the graphql schema for the frontend developer to use.

An example is shown in ent-rsvp.