Add generics to AllowIfViewerIsRule
Allows a caller to specify the Ent type into the AllowIfViewerIsRule, so that Typescript can validate the string key passed in.

This isn't backwards compatible, so leaving up to you whether you want this in. It would just require going in and adding a generic to every use case.

VALID
`new AllowIfViewerIsRule<Ent>('id')`

TYPE ERROR
`new AllowIfViewerIsRule<Ent>('is')`