kill merge revisions
fixes https://github.com/lolopinto/ent/issues/620

when there are multiple branches and we want to create a new revision/change, instead of using a merge revision like we did in https://github.com/lolopinto/ent/pull/546, we should use multiple down branches.
this makes it easier to keep track of what's going on. if one is trying to switch between branches and trying to keep track of changes between branches and has to count the merge revision, it becomes too difficult and annoying