cleanup NodeInstance in generated code
it's annoying, have had to clean it up/maintain it as we support different types
use "obj" consistently instead of trying to map it to the type of ent

this also fixes a weird bug if a separate object was defined in another file e.g. `EmailInfoType` in `contact_type.ts` where the NodeInstance value wasn't consistent because the file is passing "contact" and the code was expecting "obj"

the types are currently defined in the base ent file if they are returned from a custom field in that.

it's just sadness over time as things changed/evolved/haven't been cleaned up yet.