Add an ability to specify field aliases on a per-field basis
When using joins in `CustomClauseQuery`, if you want to sort by a column on a joined table, you need to include that column in the select fields. However, if you are using aliases, there's no way to add a column not on the main table alias. This adds a new field type beyond just `string` that lets you specify the alias for just that single field, which will override the default alias for the table.