Adjust edge queries to support multiple sort columns
When using `BaseEdgeQuery`, if you are sorting over multiple non-unique columns, the pagination was coming back incorrectly because the cursor was only storing a single sort column (plus the cursor key, usually the primary key uuid). This changes queries to allow for any number of sort columns and to run through the same SQL-generation despite the amount or type of sorting going on.

Fixes #1750 and #1751